requirements.txt
openai
streamlit
nest-asyncio
python-dotenv
